+++ R/convert_shapes_to_sf.R
@@ -7,6 +7,10 @@
 #'   If `NULL` (the default), all shapes are converted.
 #' @param crs The CRS of the resulting object, either as an EPSG code or as an
 #'   `crs` object. Defaults to 4326 (WGS 84).
+#' @param sort_sequence A logical. Whether to sort shapes by
+#'   `shape_pt_sequence`. Defaults to `FALSE`, otherwise spec-compliant feeds,
+#'   in which shape points are already ordered by `shape_pt_sequence`, would be
+#'   penalized through longer processing times.
 #'
 #' @return A `LINESTRING sf` object.
 #'
@@ -22,7 +26,10 @@
 #' shapes_sf
 #'
 #' @export
-convert_shapes_to_sf <- function(gtfs, shape_id = NULL, crs = 4326) {
+convert_shapes_to_sf <- function(gtfs,
+                                 shape_id = NULL,
+                                 crs = 4326,
+                                 sort_sequence = FALSE) {
   gtfs <- assert_and_assign_gtfs_object(gtfs)
   checkmate::assert_character(shape_id, null.ok = TRUE, any.missing = FALSE)
   checkmate::assert(
@@ -30,6 +37,7 @@ convert_shapes_to_sf <- function(gtfs, shape_id = NULL, crs = 4326) {
     checkmate::check_class(crs, "crs"),
     combine = "or"
   )
+  checkmate::assert_logical(sort_sequence, any.missing = FALSE, len = 1)
 
   gtfsio::assert_field_class(
     gtfs,
@@ -57,6 +65,13 @@ convert_shapes_to_sf <- function(gtfs, shape_id = NULL, crs = 4326) {
     shapes <- gtfs$shapes[shape_id %chin% relevant_shapes]
   } else {
     shapes <- gtfs$shapes
+  }
+
+  if (sort_sequence) {
+    gtfsio::assert_field_class(gtfs, "shapes", "shape_pt_sequence", "integer")
+
+    if (is.null(shape_id)) shapes <- data.table::copy(shapes)
+    shapes <- data.table::setorderv(shapes, c("shape_id", "shape_pt_sequence"))
   }
 
   # create an empty LINESTRING sf if shapes is empty